requirements.txt
Flask
requests
SpeechRecognition
opencv-python
google-generativeai
python-dotenv
